+++ include/ZipLocalFile.h
@@ -2,6 +2,7 @@
 #include<vector>
 #include<string>
 #include<utility>
+#include<cstddef>
 
 #ifndef ZIP_LOCAL_FILE_H_
 #define ZIP_LOCAL_FILE_H_
@@ -12,10 +13,10 @@ extern uint32_t const ZIP_LOCAL_FILE_HEADER_DATA_DESCRIPTOR_SIGNATURE;
 // Class chứa thông tin local file trong file zip 
 class ZipLocalFile {
 public:
-    ZipLocalFile(std::vector<uint8_t> const& data, int& _start_offset);
-    std::pair<int,int> GetFileName() const;
-    std::pair<int,int> GetExtraField() const;
-    std::pair<int,int> GetData() const;
+    ZipLocalFile(std::vector<uint8_t> const& data, size_t& _start_offset);
+    std::pair<size_t,size_t> GetFileName() const;
+    std::pair<size_t,size_t> GetExtraField() const;
+    std::pair<size_t,size_t> GetData() const;
     int IsEncrypted() const;
     uint16_t GetCompressionMethod() const;
     uint32_t GetCrc32() const;
@@ -35,9 +36,9 @@ protected:
     uint16_t file_name_length;
     uint16_t extra_field_length;
 
-    int file_name_start_offset, file_name_end_offset;
-    int extra_field_start_offset, extra_field_end_offset;
-    int data_start_offset, data_end_offset;
+    size_t file_name_start_offset, file_name_end_offset;
+    size_t extra_field_start_offset, extra_field_end_offset;
+    size_t data_start_offset, data_end_offset;
     std::string file_name;
 };
